refactor(current-user): Extract current user logic in own module
