+++ src/app/hooks/useApp.ts
@@ -14,6 +14,7 @@ import { PersistentStorage } from '@/common/persistent-storage';
 import { UserService } from '@/users/services';
 import { StudentsService } from '@/students/services';
 import { StudentGroupsService } from '@/student-groups/services';
+import { CurrentUserService } from '@/current-user/services';
 
 function initServices(): AppServices {
   const persistentStorage = new PersistentStorage();
@@ -27,6 +28,7 @@ function initServices(): AppServices {
 
   const authService = new AuthService(httpClient, authTokensStorage);
   const userService = new UserService(httpClient);
+  const currentUserService = new CurrentUserService(httpClient);
   const studentsService = new StudentsService(httpClient);
   const studentGroupsService = new StudentGroupsService(httpClient);
   const departmentsService = new DepartmentsService(httpClient);
@@ -37,6 +39,7 @@ function initServices(): AppServices {
   return {
     authService,
     userService,
+    currentUserService,
     studentsService,
     studentGroupsService,
     departmentsService,
