+++ src/auth/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from './useIsAuthorizedQuery.ts';
 export * from './useLogin.ts';
 export * from './useLogout.ts';
-export * from './usePasswordChange.ts';
-export * from './usePasswordReset.ts';
+export * from './useChangePassword.ts';
+export * from './useResetPassword.ts';
