+++ src/current-user/entities/CurrentUser.ts
@@ -1,0 +1,3 @@
+import { User } from '@/users/entities';
+
+export type CurrentUser = User;
