+++ src/current-user/services/dto/CurrentUserDto.ts
@@ -1,0 +1,3 @@
+import { UserDto } from '@/users/services/dto';
+
+export type CurrentUserDto = UserDto;
