+++ src/current-user/services/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './CurrentUserDto';
