--- src/users/entities/CurrentUser.ts
@@ -1,3 +1,0 @@
-import { User } from './User';
-
-export type CurrentUser = User;
