--- src/users/services/dto/CurrentUserDto.ts
@@ -1,3 +1,0 @@
-import { UserDto } from './UserDto.ts';
-
-export type CurrentUserDto = UserDto;
