+++ src/users/services/dto/index.ts
@@ -1,2 +1,1 @@
-export * from './CurrentUserDto';
 export * from './UserDto';
